cleaned up

+++ IndexBuffer.cpp
@@ -2,8 +2,8 @@
 
 
 IndexBuffer::IndexBuffer(DeviceManager& deviceManager,
-		VkCommandPool commandPool,
-		std::vector<uint32_t> indices)
+                         VkCommandPool commandPool,
+                         std::vector<uint32_t> indices)
 	:
 	indices(indices)
 {
@@ -12,17 +12,20 @@ IndexBuffer::IndexBuffer(DeviceManager& deviceManager,
 	VkBuffer stagingBuffer;
 	VkDeviceMemory stagingBufferMemory;
 	createBuffer(bufferSize, VK_BUFFER_USAGE_TRANSFER_SRC_BIT, VK_MEMORY_PROPERTY_HOST_VISIBLE_BIT |
-		VK_MEMORY_PROPERTY_HOST_COHERENT_BIT, stagingBuffer, stagingBufferMemory, deviceManager.getDevice(), deviceManager.getPhysicalDevice());
+	             VK_MEMORY_PROPERTY_HOST_COHERENT_BIT, stagingBuffer, stagingBufferMemory, deviceManager.getDevice(),
+	             deviceManager.getPhysicalDevice());
 
 	void* data;
 	vkMapMemory(deviceManager.getDevice(), stagingBufferMemory, 0, bufferSize, 0, &data);
-	memcpy(data, indices.data(), (size_t)bufferSize);
+	memcpy(data, indices.data(), bufferSize);
 	vkUnmapMemory(deviceManager.getDevice(), stagingBufferMemory);
 
 	createBuffer(bufferSize, VK_BUFFER_USAGE_TRANSFER_DST_BIT | VK_BUFFER_USAGE_INDEX_BUFFER_BIT,
-		VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT, buffer, bufferMemory, deviceManager.getDevice(), deviceManager.getPhysicalDevice());
+	             VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT, buffer, bufferMemory, deviceManager.getDevice(),
+	             deviceManager.getPhysicalDevice());
 
-	copyBuffer(stagingBuffer, buffer, bufferSize, commandPool, deviceManager.getDevice(), deviceManager.getGraphicsQueue());
+	copyBuffer(stagingBuffer, buffer, bufferSize, commandPool, deviceManager.getDevice(),
+	           deviceManager.getGraphicsQueue());
 
 	vkDestroyBuffer(deviceManager.getDevice(), stagingBuffer, nullptr);
 	vkFreeMemory(deviceManager.getDevice(), stagingBufferMemory, nullptr);
