+++ VertexBuffer.cpp
@@ -9,6 +9,7 @@ VkVertexInputBindingDescription Vertex::getBindingDescription()
 
 	return bindingDescription;
 }
+
 std::array<VkVertexInputAttributeDescription, 4> Vertex::getAttributeDescriptions()
 {
 	std::array<VkVertexInputAttributeDescription, 4> attributeDescriptions{};
@@ -43,28 +44,31 @@ bool Vertex::operator==(const Vertex& other) const
 
 
 VertexBuffer::VertexBuffer(DeviceManager& deviceManager, VkCommandPool& commandPool,
-	const std::vector<Vertex>& vertices)
+                           const std::vector<Vertex>& vertices)
 	:
 	vertices(vertices)
 {
 	VkDeviceSize bufferSize = sizeof(vertices[0]) * vertices.size();
-	
+
 
 	VkBuffer stagingBuffer;
 	VkDeviceMemory stagingBufferMemory;
 	createBuffer(bufferSize, VK_BUFFER_USAGE_TRANSFER_SRC_BIT, VK_MEMORY_PROPERTY_HOST_VISIBLE_BIT |
-		VK_MEMORY_PROPERTY_HOST_COHERENT_BIT, stagingBuffer, stagingBufferMemory, deviceManager.getDevice(), deviceManager.getPhysicalDevice());
+	             VK_MEMORY_PROPERTY_HOST_COHERENT_BIT, stagingBuffer, stagingBufferMemory, deviceManager.getDevice(),
+	             deviceManager.getPhysicalDevice());
 
 
 	void* data;
 	vkMapMemory(deviceManager.getDevice(), stagingBufferMemory, 0, bufferSize, 0, &data);
-	memcpy(data, vertices.data(), (size_t)bufferSize);
+	memcpy(data, vertices.data(), bufferSize);
 	vkUnmapMemory(deviceManager.getDevice(), stagingBufferMemory);
 
 	createBuffer(bufferSize, VK_BUFFER_USAGE_TRANSFER_DST_BIT | VK_BUFFER_USAGE_VERTEX_BUFFER_BIT,
-		VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT, buffer, bufferMemory, deviceManager.getDevice(), deviceManager.getPhysicalDevice());
+	             VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT, buffer, bufferMemory, deviceManager.getDevice(),
+	             deviceManager.getPhysicalDevice());
 
-	copyBuffer(stagingBuffer, buffer, bufferSize, commandPool, deviceManager.getDevice(), deviceManager.getGraphicsQueue());
+	copyBuffer(stagingBuffer, buffer, bufferSize, commandPool, deviceManager.getDevice(),
+	           deviceManager.getGraphicsQueue());
 
 	vkDestroyBuffer(deviceManager.getDevice(), stagingBuffer, nullptr);
 	vkFreeMemory(deviceManager.getDevice(), stagingBufferMemory, nullptr);
